+++ usvg/src/convert/text.rs
@@ -81,11 +81,13 @@ fn convert_chunks(
         let stroke = stroke::convert(tree, attrs, true);
         let decoration = conv_tspan_decoration2(tree, text_elem, &tspan);
         let visibility = super::convert_visibility(attrs);
+        let baseline_shift = conv_baseline_shift(attrs);
         let span = tree::TextSpan {
             visibility,
             fill,
             stroke,
-            font: convert_font(attrs, opt),
+            font: conv_font(attrs, opt),
+            baseline_shift,
             decoration,
             text,
         };
@@ -192,7 +194,50 @@ fn conv_text_anchor(attrs: &svgdom::Attributes) -> tree::TextAnchor {
     }
 }
 
-fn convert_font(attrs: &svgdom::Attributes, opt: &Options) -> tree::Font {
+fn conv_baseline_shift(attrs: &svgdom::Attributes) -> tree::BaselineShift {
+    let av = attrs.get_value(AId::BaselineShift);
+    match av {
+        Some(AValue::String(ref s)) => {
+            match s.as_str() {
+                "baseline" => tree::BaselineShift::Baseline,
+                "sub" => tree::BaselineShift::Subscript,
+                "super" => tree::BaselineShift::Superscript,
+                _ => {
+                    warn!("An invalid 'baseline-shift' value: '{}'. Fallback to 'baseline'.", s);
+                    tree::BaselineShift::Baseline
+                }
+            }
+        }
+        Some(AValue::Length(len)) => {
+            if len.num.is_fuzzy_zero() {
+                tree::BaselineShift::Baseline
+            } else {
+                match len.unit {
+                    Unit::Percent => {
+                        tree::BaselineShift::Percent(len.num)
+                    }
+                    _ => {
+                        warn!("'baseline-shift' value must be a number or a percent.");
+                        tree::BaselineShift::Baseline
+                    }
+                }
+            }
+        }
+        Some(AValue::Number(n)) => {
+            tree::BaselineShift::Number(*n)
+        }
+        None => {
+            // Fallback to a default value.
+            tree::BaselineShift::Baseline
+        }
+        _ => {
+            warn!("An invalid 'baseline-shift' value: '{:?}'. Fallback to 'baseline'.", av);
+            tree::BaselineShift::Baseline
+        }
+    }
+}
+
+fn conv_font(attrs: &svgdom::Attributes, opt: &Options) -> tree::Font {
     let style = attrs.get_str_or(AId::FontStyle, "normal");
     let style = match style {
         "normal" => tree::FontStyle::Normal,