+++ usvg/src/tree/convert.rs
@@ -363,6 +363,7 @@ fn conv_elements(
                         conv_fill(tree, &tspan.fill, defs, parent, &mut tspan_elem);
                         conv_stroke(tree, &tspan.stroke, defs, &mut tspan_elem);
                         conv_font(&tspan.font, &mut tspan_elem);
+                        conv_baseline_shift(tspan.baseline_shift, &mut tspan_elem);
 
                         if tspan.text.contains("  ") {
                             is_preserve_required = true;
@@ -635,6 +636,21 @@ fn conv_font(
             FontStretch::UltraExpanded => "ultra-expanded",
         }
     ));
+}
+
+fn conv_baseline_shift(
+    baseline_shift: BaselineShift,
+    node: &mut svgdom::Node,
+) {
+    let av: AValue = match baseline_shift {
+        BaselineShift::Baseline => "baseline".into(),
+        BaselineShift::Subscript => "sub".into(),
+        BaselineShift::Superscript => "super".into(),
+        BaselineShift::Percent(n) => svgdom::Length::new(n, svgdom::LengthUnit::Percent).into(),
+        BaselineShift::Number(n) => n.into(),
+    };
+
+    node.set_attribute((AId::BaselineShift, av));
 }
 
 fn conv_link(